- добавлено удаление  области холста (не 6912) - доработана возможность отменять/вернуть действия для 6912 (для атрибутов есть проблемы объеденения, что вызывает неточности)

+++ Source/Window/Sprite/UndoAction.h
@@ -3,6 +3,6 @@
 struct FPixelToCanvas
 {
 	std::vector<ImVec2> Position;
-	std::vector<uint8_t> Color;
+	std::vector<uint32_t> Color;
 	uint32_t Canvas;
 };
